reworking bot builder
